// src/app/interfaces/common/supplier.interface.ts
export interface Supplier {
  _id?: string;
  shop?: any;
  name?: string;
  email?: string;
  phone?: string;
  country?: string;
  city?: string;
  address?: string;
  description?: string;
  image?: string;
  totalDue?: number;
  totalPaid?: number;
  alternatePhone?: string;
  contactPerson?: string;
  taxId?: string;
  notes?: string;
  createdAtString?: string;
  updatedAtString?: string;
  createdAt?: Date;
  updatedAt?: Date;
}